Проверка блочной записи

+++ src/addfiles2arch.h
@@ -2,6 +2,7 @@
 #define ADDFILES2ARCH_H
 
 #include "logging.h"
+#include <inttypes.h>
 #include <stdio.h>
 
 #define BUF_LEN (8)
@@ -9,6 +10,6 @@
 
 #define IO_WRITE_ERROR (2451)
 
-int writeBytes(FILE *f, unsigned char *buf, int64_t k_bytes);
+int writeBytes(FILE *f, unsigned char *buf, size_t k_bytes);
 
 #endif // ADDFILES2ARCH_H
